Small corrections of minor mistakes / tidiness.
Frontend (src/js/shipping-workshop-block/block.js):

**frontend-step-02-extra-credit-1** suggests to use debounced function for setExtensionData,
though I observed the code uses the normal one. The funny thing, that I changed it to the debounced
one and it worked couple of times, and then started throwing an error for 'extensions' parameter.
Would be happy to know what could go wrong with debouncing there. Would agree that for the select field
it is not so necessary, one can't change it so fast.

**frontend-step-02-extra-credit-2** useEffect dependencies don't contain hasInteractedWithOtherInput, so the error
is not displayed immediately onBlur.

**frontend-step-03-extra-credit** is also about the debounce. The code contains the simple function, while useEffect
dependencies contain the debounced function. It works fine with debounce.

Edited spoilers for frontend steps 04-extra-credit, 05 and 07-extra-credit to match tasks in the code.
(spoiler for 05 contained what should be done in 07-extra-credit (practically spoilers were the
same, with 05 being ahead of time), while 04-extra-credit seemed to be what should be in 05)

Backend:

**shipping-workshop-extend-store-endpoint.php**
use Automattic\WooCommerce\Blocks\Package and use Automattic\WooCommerce\Blocks\StoreApi\Schemas\CartSchema were never called.

**shipping-workshop-extend-woo-core.php** none of three use was called in the code.

**shipping-workshop-extend-woo-core.php** contained a never called function extend_checkout_schema(),
which is placed in shipping-workshop-extend-store-endpoint.php and called from there.

Updated dependencies to the latest versions and added overrides for vulnerable @wordpress/scripts dependencies.